requirements.txt
dnspython==1.15.0
pykerberos==1.2.1
requests==2.18.4
requests-kerberos==0.12.0
